refactor(checker): split annotation-text helpers out of assignment_formatting.rs to hold 2000 LOC
Move the 4 assignment target/source annotation-text methods
(direct_assignment_target_annotation_node/_text,
source_assignment_target_annotation_text, annotation_text_from_colon_fragment)
into a new sibling diagnostic_source/assignment_annotation_text.rs (pure code
motion, pub(super)) so assignment_formatting.rs stays under the 2000-line arch
cap (was 2042 after this PR's aliasSymbol-retention display work). No behavior
change; all deps used are already pub(crate)/pub(in crate::error_reporter).
